Ensure that resources are finalized in dagstermill
Summary: This picks up on Uma's work in https://github.com/dagster-io/dagster/pull/895. It is a bit
of a bear. The atexit-based approach in that PR does not work, nor does hooking into ipython's
event system. Finding the right place in the papermill machinery to execute this isn't obvious,
and the underlying nbconvert concept of a "preprocessor" is very misleading in the context of
notebook execution.

Test Plan: Unit test

Reviewers: schrockn, natekupp

Subscribers: