EnvironmentSchema in GraphQL
Summary: Right now the way you get at the config schema of a pipeline is
as a property off of that pipeline. I think this is bit of an awkward
way to model it as each of those have to take mode and solidSubset and
then you have to have an error union in a nested part of the tree.

Instead here we just have a top level environment schema that we can
hang elements off of. Interested in hearing feedback on this. I'm not
100% convinced that it is better.

Test Plan: unit tests

Reviewers: max, alangenfeld, bengotow

Differential Revision: https://dagster.phacility.com/D201