Improve airline_demo init time
Summary: So this is interesting. Papermill/scrapbook has a pretty hefty startup time cost.

It looks like it starts the IPython interpreter on import which is about 2 seconds. Not sure
if this is worth getting in tiff about but it is annoying. Had to do shenanigans to get this working

Test Plan: Use profiler. See improvement.

Reviewers: max, natekupp