(airline-demo-testing 10/N) Add single solid test for process_q2_data
Summary: This was an interesting exercise. It was definitely a process to get this
up and running

Things to note:

1) Generate test data. I did this by executing the pipeline and the configuring
ingest_{april,may,june,master_cord}_on_time_data to output a CSV during a run
via the configuration system. I then moved these files into the unit_test directory

2) I have to rely on spark's memoization of spark session create in order to get this
to work. If it didn't have this capability, it would be difficult to construct the
input dataframes on the same context. Would have to create a solid solely for the
test case that did this and then create an emphemeral pipeline.

3) The lack of scoping for resources hurt a bit here, because I had to fake up config.

4) This uses snapshots which I think are an excellent testing methodology for data
pipelines which I do not believe are in widespread use in the community.

5) Overall this was still quite a bit of work. Dagit worked beautifully but this took
quite a bit of setup.

The alternative would be to create a special purpose pipeline for this test case.

Test Plan: Unit tests

Reviewers: max, natekupp