RFC (not for commit): DSL for dependency graph construction
Summary: Alright this is a fairly big proposal, but I think it is a huge
step function in usability for the system and would propose that we
block shipping 0.5.0 on it or something like it. Instead of having to
construct dictionaries we use python API/DSL to construct the DAG
structure. In addition we introduce the @pipeline decorator
in order to to do this.

This pretty dramatically reduces boilerplate (both in that the
list of solids is now inferred from the dep graph, the SolidInstance
crap is gone, and the code is easier to reason about and structure
that a dictionary). Just look at the airline demo in this PR for an
example.

Natural extensions would be to also add a very similar decorator for
composites. I would propose calling it @graph. So we would have
arbitrary graphs of computation that can compose and call eachother.

A pipeline would be a graph that is "bound" to set of presets and modes.
Pre-defined entry points.

I think the before/after picture for a complicated airflow dag expressed
on this would be absolutely epic.

Test Plan: Unit tests. Run in dagit

Differential Revision: https://dagster.phacility.com/D275

Reviewers: natekupp, alangenfeld, max